Issue #982. Add a LOCALHOST-only /test-files/ route, and a window-helpers.js.
Before, when Intern was injecting the code to postMessage blobs, Firefox
was throwing an x-ray wrapper permission error, which killed the test.

Presumably this is related to how WebDriver/Intern injects the JS test files.
As a test-only workaround, make it possible to load a test helper file so
methods can be called trusted content.